# docs/openapi.yaml
openapi: 3.0.3
info:
  title: MICO360 Tasks API
  version: 1.0.0
  description: >-
    REST API for MICO360 Tasks (web app, Chrome extension, and the Phase 2 Android
    app). All responses use a JSON envelope: success is `{ "data": ... }`, errors
    are `{ "error": { "code": string, "message": string, "details"?: any } }`.
    Authenticate by sending `Authorization: Bearer <accessToken>`.
servers:
  - url: http://localhost:4000/api/v1
    description: Local development
  - url: https://task.mico360.com/api/v1
    description: Production
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Task:
      type: object
      properties:
        id: { type: string }
        key: { type: string, example: MICO-142 }
        title: { type: string }
        description: { type: string, nullable: true }
        projectId: { type: string }
        columnId: { type: string }
        position: { type: integer }
        priority: { type: string, enum: [LOW, NORMAL, HIGH, URGENT] }
        dueDate: { type: string, format: date-time, nullable: true }
        progress: { type: integer }
        recurrenceRule:
          nullable: true
          description: The repeat rule. Only the newest copy of a recurring series carries it.
          $ref: '#/components/schemas/RecurrenceRule'
        recurrenceParentId:
          type: string
          nullable: true
          description: Set on every copy of a recurring series (the first task's id; the first task links to itself once it has a copy).
        recurrenceNextId:
          type: string
          nullable: true
          description: >
            GET /tasks/{id} only. The copy made from this task. Set means this is an earlier copy: setting
            a repeat on it is refused with 409 RECURRENCE_NOT_NEWEST (it would start a second series).
    RecurrenceRule:
      type: object
      description: >
        A task's repeat schedule. When the next copy is made it keeps the title, description,
        priority, estimate, assignees, watchers, tags and checklist (unticked), lands in the
        project's first open column, and takes over the rule. A task never gets two next copies
        (enforced by the database), so completing it twice, or from several apps at once, makes one.
      required: [freq, interval]
      properties:
        freq: { type: string, enum: [DAILY, WEEKLY, MONTHLY, QUARTERLY, YEARLY] }
        interval: { type: integer, minimum: 1, description: 'Every N days / weeks / months / quarters / years.' }
        count: { type: integer, nullable: true, description: 'End after this many tasks in the series.' }
        until: { type: string, nullable: true, description: 'Last date (YYYY-MM-DD) a copy may be due.' }
        weekdays: { type: array, items: { type: integer, minimum: 0, maximum: 6 }, description: 'WEEKLY: the days (0 = Sunday).' }
        dayOfMonth: { type: integer, minimum: 1, maximum: 31, description: 'MONTHLY / QUARTERLY: the day (31 = the last day of each month).' }
        nthWeekday:
          type: object
          description: 'MONTHLY / QUARTERLY, instead of dayOfMonth: "the 2nd Tuesday" or "the last Friday".'
          properties:
            week: { type: integer, enum: [1, 2, 3, 4, -1], description: '-1 = the last one of the month.' }
            day: { type: integer, minimum: 0, maximum: 6 }
        createNext:
          type: string
          enum: [ON_COMPLETE, ON_SCHEDULE]
          description: >
            ON_COMPLETE (default): the next copy is made when this one is completed, dated by the
            rule and never already overdue (missed dates are skipped). ON_SCHEDULE: a copy is made
            on each due date whether or not the previous one is done (an hourly server job; missed
            days collapse into one copy).
        paused: { type: boolean, description: 'No new copies while paused.' }
...
security:
  - bearerAuth: []
paths:
...
  /tasks:
    get:
      summary: List tasks (optionally filtered by projectId/columnId)
      parameters:
        - { name: projectId, in: query, schema: { type: string } }
        - { name: columnId, in: query, schema: { type: string } }
      responses: { '200': { description: OK } }
    post:
      summary: Create a task
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/Task' } } }
      responses: { '201': { description: Created } }
...
  /tasks/{id}:
    get:
      summary: Get a task
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      responses: { '200': { description: OK }, '404': { description: Not found } }
    put:
      summary: Update a task
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      responses: { '200': { description: OK } }
    delete:
      summary: Delete a task (admin)
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      responses: { '204': { description: Deleted } }
  /tasks/{id}/move:
    patch:
      summary: Move a task to a column (Kanban drag-drop persistence)
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [columnId], properties: { columnId: { type: string }, position: { type: integer } } } } }
      responses: { '200': { description: OK } }
...